+++ packages/gator-permissions-controller/src/decodePermission/rules/makePermissionRule.test.ts
@@ -30,8 +30,8 @@ describe('makePermissionRule', () => {
       {
         enforcer: timestampEnforcer,
         terms: createTimestampTerms({
-          timestampAfterThreshold: 0,
-          timestampBeforeThreshold: 1720000,
+          afterThreshold: 0,
+          beforeThreshold: 1720000,
         }),
         args: '0x' as Hex,
       },
@@ -140,8 +140,8 @@ describe('makePermissionRule', () => {
       {
         enforcer: timestampEnforcer,
         terms: createTimestampTerms({
-          timestampAfterThreshold: 0,
-          timestampBeforeThreshold: 1720000,
+          afterThreshold: 0,
+          beforeThreshold: 1720000,
         }),
         args: '0x' as Hex,
       },
@@ -178,8 +178,8 @@ describe('makePermissionRule', () => {
       {
         enforcer: timestampEnforcer,
         terms: createTimestampTerms({
-          timestampAfterThreshold: 0,
-          timestampBeforeThreshold: 1720000,
+          afterThreshold: 0,
+          beforeThreshold: 1720000,
         }),
         args: '0x' as Hex,
       },